requirements.txt
flask==3.0.2
python-dotenv==1.0.1
langchain[all]==0.2.5
langchain-community==0.2.5
langchain-pinecone==0.1.1
pinecone-client==3.2.2
langchain-openai==0.1.8
chardet==5.2.0
boto3==1.34.131
pdfminer.six==20221105
docx2txt==0.8
python-docx==1.1.2
PyMuPDF==1.24.5
openai==1.35.3